moved api to a separate folder in root and added npm scripts
