+++ api/config.js
@@ -1,18 +1,10 @@
 module.exports = {
   development: {
     isProduction: false,
-    port: process.env.PORT,
     apiPort: process.env.APIPORT,
-    app: {
-      name: 'React Redux Example Development'
-    }
   },
   production: {
     isProduction: true,
-    port: process.env.PORT,
     apiPort: process.env.APIPORT,
-    app: {
-      name: 'React Redux Example Production'
-    }
   }
 }[process.env.NODE_ENV || 'development'];
