+++ bin/api.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+require('../server.babel'); // babel registration (runtime transpilation for node)
+require('../api/api');
